+++ src/SharpVk/DescriptorSetLayoutBinding.gen.cs
@@ -78,7 +78,7 @@ namespace SharpVk
             if (this.ImmutableSamplers != null)
             {
                 var fieldPointer = (Interop.Sampler*)(Interop.HeapUtil.AllocateAndClear<Interop.Sampler>(this.ImmutableSamplers.Length).ToPointer());
-                for(int index = 0; index < this.ImmutableSamplers.Length; index++)
+                for(int index = 0; index < (uint)(this.ImmutableSamplers.Length); index++)
                 {
                     fieldPointer[index] = this.ImmutableSamplers[index].handle;
                 }
