+++ src/SharpVk/DescriptorSetLayoutCreateInfo.gen.cs
@@ -60,7 +60,7 @@ namespace SharpVk
             if (this.Bindings != null)
             {
                 var fieldPointer = (Interop.DescriptorSetLayoutBinding*)(Interop.HeapUtil.AllocateAndClear<Interop.DescriptorSetLayoutBinding>(this.Bindings.Length).ToPointer());
-                for(int index = 0; index < this.Bindings.Length; index++)
+                for(int index = 0; index < (uint)(this.Bindings.Length); index++)
                 {
                     this.Bindings[index].MarshalTo(&fieldPointer[index]);
                 }
