+++ src/SharpVk/FramebufferCreateInfo.gen.cs
@@ -97,7 +97,7 @@ namespace SharpVk
             if (this.Attachments != null)
             {
                 var fieldPointer = (Interop.ImageView*)(Interop.HeapUtil.AllocateAndClear<Interop.ImageView>(this.Attachments.Length).ToPointer());
-                for(int index = 0; index < this.Attachments.Length; index++)
+                for(int index = 0; index < (uint)(this.Attachments.Length); index++)
                 {
                     fieldPointer[index] = this.Attachments[index].handle;
                 }
