+++ src/SharpVk/Khr/DescriptorUpdateTemplateCreateInfo.gen.cs
@@ -105,7 +105,7 @@ namespace SharpVk.Khr
             if (this.DescriptorUpdateEntries != null)
             {
                 var fieldPointer = (SharpVk.Khr.DescriptorUpdateTemplateEntry*)(Interop.HeapUtil.AllocateAndClear<SharpVk.Khr.DescriptorUpdateTemplateEntry>(this.DescriptorUpdateEntries.Length).ToPointer());
-                for(int index = 0; index < this.DescriptorUpdateEntries.Length; index++)
+                for(int index = 0; index < (uint)(this.DescriptorUpdateEntries.Length); index++)
                 {
                     fieldPointer[index] = this.DescriptorUpdateEntries[index];
                 }
