+++ src/SharpVk/Khr/PresentInfo.gen.cs
@@ -77,7 +77,7 @@ namespace SharpVk.Khr
             if (this.WaitSemaphores != null)
             {
                 var fieldPointer = (Interop.Semaphore*)(Interop.HeapUtil.AllocateAndClear<Interop.Semaphore>(this.WaitSemaphores.Length).ToPointer());
-                for(int index = 0; index < this.WaitSemaphores.Length; index++)
+                for(int index = 0; index < (uint)(this.WaitSemaphores.Length); index++)
                 {
                     fieldPointer[index] = this.WaitSemaphores[index].handle;
                 }
@@ -91,7 +91,7 @@ namespace SharpVk.Khr
             if (this.Swapchains != null)
             {
                 var fieldPointer = (Interop.Khr.Swapchain*)(Interop.HeapUtil.AllocateAndClear<Interop.Khr.Swapchain>(this.Swapchains.Length).ToPointer());
-                for(int index = 0; index < this.Swapchains.Length; index++)
+                for(int index = 0; index < (uint)(this.Swapchains.Length); index++)
                 {
                     fieldPointer[index] = this.Swapchains[index].handle;
                 }
@@ -104,7 +104,7 @@ namespace SharpVk.Khr
             if (this.ImageIndices != null)
             {
                 var fieldPointer = (uint*)(Interop.HeapUtil.AllocateAndClear<uint>(this.ImageIndices.Length).ToPointer());
-                for(int index = 0; index < this.ImageIndices.Length; index++)
+                for(int index = 0; index < (uint)(this.ImageIndices.Length); index++)
                 {
                     fieldPointer[index] = this.ImageIndices[index];
                 }
@@ -117,7 +117,7 @@ namespace SharpVk.Khr
             if (this.Results != null)
             {
                 var fieldPointer = (Result*)(Interop.HeapUtil.AllocateAndClear<Result>(this.Results.Length).ToPointer());
-                for(int index = 0; index < this.Results.Length; index++)
+                for(int index = 0; index < (uint)(this.Results.Length); index++)
                 {
                     fieldPointer[index] = this.Results[index];
                 }
