+++ src/SharpVk/Khx/BindImageMemoryInfo.gen.cs
@@ -89,7 +89,7 @@ namespace SharpVk.Khx
             if (this.DeviceIndices != null)
             {
                 var fieldPointer = (uint*)(Interop.HeapUtil.AllocateAndClear<uint>(this.DeviceIndices.Length).ToPointer());
-                for(int index = 0; index < this.DeviceIndices.Length; index++)
+                for(int index = 0; index < (uint)(this.DeviceIndices.Length); index++)
                 {
                     fieldPointer[index] = this.DeviceIndices[index];
                 }
@@ -103,7 +103,7 @@ namespace SharpVk.Khx
             if (this.SFRRects != null)
             {
                 var fieldPointer = (Rect2D*)(Interop.HeapUtil.AllocateAndClear<Rect2D>(this.SFRRects.Length).ToPointer());
-                for(int index = 0; index < this.SFRRects.Length; index++)
+                for(int index = 0; index < (uint)(this.SFRRects.Length); index++)
                 {
                     fieldPointer[index] = this.SFRRects[index];
                 }
