+++ src/SharpVk/Khx/DeviceGroupDeviceCreateInfo.gen.cs
@@ -50,7 +50,7 @@ namespace SharpVk.Khx
             if (this.PhysicalDevices != null)
             {
                 var fieldPointer = (Interop.PhysicalDevice*)(Interop.HeapUtil.AllocateAndClear<Interop.PhysicalDevice>(this.PhysicalDevices.Length).ToPointer());
-                for(int index = 0; index < this.PhysicalDevices.Length; index++)
+                for(int index = 0; index < (uint)(this.PhysicalDevices.Length); index++)
                 {
                     fieldPointer[index] = this.PhysicalDevices[index].handle;
                 }
