+++ src/SharpVk/Nv/Win32KeyedMutexAcquireReleaseInfo.gen.cs
@@ -86,7 +86,7 @@ namespace SharpVk.Nv
             if (this.AcquireSyncs != null)
             {
                 var fieldPointer = (Interop.DeviceMemory*)(Interop.HeapUtil.AllocateAndClear<Interop.DeviceMemory>(this.AcquireSyncs.Length).ToPointer());
-                for(int index = 0; index < this.AcquireSyncs.Length; index++)
+                for(int index = 0; index < (uint)(this.AcquireSyncs.Length); index++)
                 {
                     fieldPointer[index] = this.AcquireSyncs[index].handle;
                 }
@@ -99,7 +99,7 @@ namespace SharpVk.Nv
             if (this.AcquireKeys != null)
             {
                 var fieldPointer = (ulong*)(Interop.HeapUtil.AllocateAndClear<ulong>(this.AcquireKeys.Length).ToPointer());
-                for(int index = 0; index < this.AcquireKeys.Length; index++)
+                for(int index = 0; index < (uint)(this.AcquireKeys.Length); index++)
                 {
                     fieldPointer[index] = this.AcquireKeys[index];
                 }
@@ -112,7 +112,7 @@ namespace SharpVk.Nv
             if (this.AcquireTimeoutMilliseconds != null)
             {
                 var fieldPointer = (uint*)(Interop.HeapUtil.AllocateAndClear<uint>(this.AcquireTimeoutMilliseconds.Length).ToPointer());
-                for(int index = 0; index < this.AcquireTimeoutMilliseconds.Length; index++)
+                for(int index = 0; index < (uint)(this.AcquireTimeoutMilliseconds.Length); index++)
                 {
                     fieldPointer[index] = this.AcquireTimeoutMilliseconds[index];
                 }
@@ -126,7 +126,7 @@ namespace SharpVk.Nv
             if (this.ReleaseSyncs != null)
             {
                 var fieldPointer = (Interop.DeviceMemory*)(Interop.HeapUtil.AllocateAndClear<Interop.DeviceMemory>(this.ReleaseSyncs.Length).ToPointer());
-                for(int index = 0; index < this.ReleaseSyncs.Length; index++)
+                for(int index = 0; index < (uint)(this.ReleaseSyncs.Length); index++)
                 {
                     fieldPointer[index] = this.ReleaseSyncs[index].handle;
                 }
@@ -139,7 +139,7 @@ namespace SharpVk.Nv
             if (this.ReleaseKeys != null)
             {
                 var fieldPointer = (ulong*)(Interop.HeapUtil.AllocateAndClear<ulong>(this.ReleaseKeys.Length).ToPointer());
-                for(int index = 0; index < this.ReleaseKeys.Length; index++)
+                for(int index = 0; index < (uint)(this.ReleaseKeys.Length); index++)
                 {
                     fieldPointer[index] = this.ReleaseKeys[index];
                 }
